+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # main
+* Delete uploaded media when deleting the post that references them
 * Bugfix: Set displayname and profile photo on the server at setup time
 * Bugfix: Enable taking new photo for your profile
 * Bugfix: Enable posting new photo from the camera